- JPA repository, dependency injection, database seeding

+++ src/main/java/com/barros/gestao_de_treinos/entities/Usuario.java
@@ -15,7 +15,7 @@ public class Usuario implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
     private String email;
     private String senhaHash;
@@ -25,7 +25,7 @@ public class Usuario implements Serializable {
     public Usuario() {
     }
 
-    public Usuario(Integer id, String nome, String email, String senhaHash, LocalDate dataNascimento,
+    public Usuario(Long id, String nome, String email, String senhaHash, LocalDate dataNascimento,
             LocalDateTime dataCriacao) {
         this.id = id;
         this.nome = nome;
@@ -35,11 +35,11 @@ public class Usuario implements Serializable {
         this.dataCriacao = dataCriacao;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
